Store data events (#7)
This PR adds storage for logging events. Currently the file is saved at the end of the experiment to the persistent data folder. On my computer this is: `'/Users/reinder.vosdewael/Library/Application Support/DefaultCompany/train/testFile.tsv'`, but this will vary depending on your username and OS. In a future PR this save location should be changed to be user defined / use a share button.

Specifically, the following changes are made: 
- The data capture system and utils are moved to the common folder.
- A frame number and timestamp have been added to the DataCaptureSystem
- A ObjectHash class has been added for creating a SHA256 hash from any object. This is currently used for creating a start/end hash of each room to confirm identicalness across runs.
- A `Movable` class was added for objects that can move. This class logs the new position whenever its position changes.
- The `Clickable` class now inherits from the `Movable` class, and logs whenever it is clicked.
- Various docstrings have been added. 
- Utils.SimpleTimer appears unused and has been removed.

Out of scope for this PR:
- 'Smart logging' events that depend on object interactions e.g. train occluded.